search-service/requirements.txt
fastapi
uvicorn
pymongo
sentence-transformers
torch
protobuf>=5.26.1
grpcio>=1.62.0
grpcio-tools>=1.62.0